+++ Cargo.toml
@@ -1,27 +1,32 @@
-[package]
-name = "novus"
-version = "0.1.0"
-authors = ["ZaphodElevated <pranav.rdoshi@gmail.com>"]
-edition = "2018"
+# [package]
+# name = "novus"
+# version = "0.1.0"
+# authors = ["ZaphodElevated <pranav.rdoshi@gmail.com>"]
+# edition = "2018"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+# # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
-[dependencies]
-serde_json = "1.0.64"
-reqwest = { version = "0.11.3", features = ["stream", "blocking"] }
-serde = { version = "1.0.125", features = ["derive"] }
-colored = "2.0.0"
-indicatif="0.16.0"
-tokio = { version = "1.5.0", features = ["full"] }
-futures-util = "0.3.14"
-ansi_term = "0.12.1"
-futures = "0.3"
-ctrlc = { version = "3.0", features = ["termination"] }
-difflib = "0.4.0"
-sha2 = "0.9.5"
-regex = "1.5.4"
-clokwerk = "0.3.5"
+# [dependencies]
+# serde_json = "1.0.64"
+# reqwest = { version = "0.11.3", features = ["stream", "blocking"] }
+# serde = { version = "1.0.125", features = ["derive"] }
+# colored = "2.0.0"
+# indicatif="0.16.0"
+# tokio = { version = "1.5.0", features = ["full"] }
+# futures-util = "0.3.14"
+# ansi_term = "0.12.1"
+# futures = "0.3"
+# ctrlc = { version = "3.0", features = ["termination"] }
+# difflib = "0.4.0"
+# sha2 = "0.9.5"
+# regex = "1.5.4"
+# clokwerk = "0.3.5"
 
-[target.'cfg(windows)'.dependencies]
-winreg = "0.9.0"
-winapi = { version = "0.3", features = ["winuser"] }
+# [target.'cfg(windows)'.dependencies]
+# winreg = "0.9.0"
+# winapi = { version = "0.3", features = ["winuser"] }
+[workspace]
+members = [
+	"cli",
+	"utils"
+]
