--- src/classes/installed_packages.rs
@@ -1,7 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-#[derive(Deserialize, Serialize, Debug, Clone)]
-pub struct Packages {
-  pub packages: Vec<String>,
-}
-
