--- src/commands/mod.rs
@@ -1,8 +1,0 @@
-pub mod clean;
-pub mod info;
-pub mod install;
-pub mod list;
-pub mod quit;
-pub mod search;
-pub mod uninstall;
-pub mod update;
