--- src/constants/mod.rs
@@ -1,3 +1,0 @@
-pub mod commands;
-pub mod help_menu;
-pub mod version;
