--- src/constants/version.rs
@@ -1,1 +1,0 @@
-pub const __VERSION__: &str = "v1.0.0";
