--- src/utils/handle_error.rs
@@ -1,6 +1,0 @@
-use colored::Colorize;
-
-pub fn handle_error_and_exit(e: String) -> ! {
-    println!("{}{}", "error ".bright_red(), e);
-    std::process::exit(0);
-}
