--- src/utils/mod.rs
@@ -1,8 +1,0 @@
-pub mod cache;
-pub mod checksum;
-pub mod display_help;
-pub mod get_package;
-pub mod handle_args;
-pub mod handle_error;
-pub mod registry;
-pub mod autoelevate;
