+++ utils/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "utils"
+version = "0.1.0"
+authors = ["Varun Potti <potti.varun07@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+colored = "2.0.0"
+winreg = "0.9.0"
+winapi = { version = "0.3", features = ["winuser"] }
+difflib = "0.4.0"
+chttp = "0.5.5"
+serde = { version = "1.0.125", features = ["derive"] }
+serde_json = "1.0.64"
+sha2 = "0.9.5"
+reqwest = { version = "0.11.3", features = ["stream", "blocking"] }
