+++ utils/src/classes/installed_packages.rs
@@ -1,0 +1,6 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Deserialize, Serialize, Debug, Clone)]
+pub struct Packages {
+  pub packages: Vec<String>,
+}
