+++ utils/src/classes/mod.rs
@@ -1,0 +1,4 @@
+pub mod installed_packages;
+pub mod local_package_info;
+pub mod package;
+pub mod package_info;
