+++ utils/src/constants/help_menu.rs
