+++ utils/src/constants/mod.rs
@@ -1,0 +1,3 @@
+pub mod commands;
+pub mod help_menu;
+pub mod version;
