+++ utils/src/constants/version.rs
@@ -1,0 +1,1 @@
+pub const __VERSION__: &str = "v1.0.0";
