+++ utils/src/lib.rs
@@ -1,0 +1,10 @@
+pub mod autoelevate;
+pub mod cache;
+pub mod checksum;
+pub mod constants;
+pub mod display_help;
+pub mod get_package;
+pub mod handle_args;
+pub mod handle_error;
+pub mod registry;
+pub mod classes;
